spake: switch to using a hash
For each group, specify an associated hash function and use it for
transcript hashing and key derivation. In key derivation, use
KRB-FX-CF2 with the initial reply key to avoid the possibility of
reducing its entropy. Since the hash for the transcript now depends on
the group, update the transcript hash with the support and challenge
messages concatenated together, and don't include a rejected
optimistic challenge in the transcript. Update test vectors
accordingly. As the PRF+ output is now used in the key derivation
hash, include it in the test vectors along with the encoded reduced
multiplier. Add a test vector (using a fictitious group) for when
multiple hash blocks are needed in key derivation.

Since we are no longer computing checksums, drop down to one assigned
key usage number, named just KEY_USAGE_SPAKE.

Also, group and enctype numbers are signed, so specify two's
complement rather than unsigned when encoding them in four-byte
binary.